move pvfactors tests
